ADD task_incremental_benchmark generator
