+++ tests/benchmarks/scenarios/test_experience_decorators.py
@@ -2,7 +2,8 @@ import unittest
 
 import torch
 
-from avalanche.benchmarks import benchmark_from_datasets, EagerCLStream, DatasetExperience, CLScenario
+from avalanche.benchmarks import benchmark_from_datasets, EagerCLStream, CLScenario
+from avalanche.benchmarks.scenarios.deprecated import DatasetExperience
 from avalanche.benchmarks.scenarios.task_aware import with_task_labels
 from avalanche.benchmarks.scenarios.supervised import with_classes_timeline
 from avalanche.benchmarks.utils import make_tensor_classification_dataset
